// src/app/models/weather-videos.ts
export interface WeatherVideos {
    Clear: string;
    Clouds: string;
    Rain: string;
    Drizzle: string;
    Thunderstorm: string;
    Snow: string;
    Mist: string;
    Smoke: string;
    Haze: string;
    Dust: string;
    Fog: string;
    Sand: string;
    Ash: string;
    Squall: string;
    Tornado: string;
    [key: string]: string;
}